core/engine/codec_engine.py
"""
编解码引擎 (FFmpeg/PyNvVideoCodec封装)
"""
import os
from typing import Dict, Any, List, Optional
import numpy as np
import ffmpeg
import logging


class EncodeResult:
    """编码结果"""
    def __init__(self, success: bool, output_path: str = "", message: str = ""):
        self.success = success
        self.output_path = output_path
        self.message = message


class CodecEngine:
    """编解码引擎"""
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
    
    def compress_video(self, input_path: str, output_path: str, config: Dict[str, Any]) -> EncodeResult:
        """视频压缩核心方法"""
        try:
            # 获取编码参数
            codec = config.get('codec', 'h264')
            preset = config.get('preset', 'medium')
            crf = config.get('crf', 23)
            bitrate = config.get('bitrate')
            framerate = config.get('framerate')
            resolution = config.get('resolution')
            hardware_acceleration = config.get('hardware_acceleration', False)
            
            # 构建FFmpeg命令
            stream = ffmpeg.input(input_path)
            
            # 设置输出参数
            output_kwargs = {
                'vcodec': codec,
                'preset': preset,
                'crf': crf,
            }
            
            # 添加比特率参数
            if bitrate:
                output_kwargs['video_bitrate'] = f'{bitrate}k'
            
            # 添加帧率参数
            if framerate:
                output_kwargs['r'] = framerate
                
            # 添加分辨率参数
            if resolution:
                output_kwargs['s'] = f'{resolution[0]}x{resolution[1]}'
            
            # 硬件加速
            if hardware_acceleration:
                if 'h264' in codec:
                    output_kwargs['vcodec'] = 'h264_nvenc'  # 示例：NVIDIA硬件编码
            
            stream = ffmpeg.output(stream, output_path, **output_kwargs)
            
            # 运行FFmpeg命令
            ffmpeg.run(stream, overwrite_output=True, quiet=True)
            
            return EncodeResult(success=True, output_path=output_path)
            
        except Exception as e:
            self.logger.error(f"视频压缩失败: {str(e)}")
            return EncodeResult(success=False, message=str(e))
    
    def extract_frames(self, video_path: str, interval: float = 1.0) -> List[np.ndarray]:
        """提取视频帧用于预览"""
        frames = []
        try:
            # 计算每隔interval秒提取一帧
            stream = ffmpeg.input(video_path, ss=0)
            stream = ffmpeg.filter(stream, 'select', f'not(mod(n,{int(interval * 30)}))')  # 假设30fps
            stream = ffmpeg.output(stream, 'pipe:', format='rawvideo', pix_fmt='rgb24')
            
            out, _ = ffmpeg.run(stream, capture_stdout=True, quiet=True)
            # 这里简化处理，实际需要根据视频尺寸解析RGB数据
            # 实际实现需要知道视频尺寸等信息
            
        except Exception as e:
            self.logger.error(f"提取视频帧失败: {str(e)}")
        
        return frames
    
    def get_hardware_acceleration_info(self) -> Dict[str, Any]:
        """获取可用硬件加速信息"""
        # 这里返回可用的硬件加速信息
        # 实际实现需要检测GPU类型和支持的编码器
        return {
            'nvenc_supported': False,  # NVIDIA NVENC支持
            'qsv_supported': False,    # Intel Quick Sync Video支持
            'vaapi_supported': False,  # VA-API支持
            'cuda_devices': [],        # CUDA设备列表
        }