requirements.txt
# GUI框架
PySide6>=6.5.0
# 可选: PyQt6>=6.5.0

# 高性能渲染
pyqtgraph>=0.13.0

# 图表可视化
matplotlib>=3.7.0

# 媒体处理
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# 视频编解码
ffmpeg-python>=0.2.0
# 可选GPU加速 (NVIDIA GPU编码)
# pynvcodec>=0.0.1  # NVIDIA GPU编码

# 工具库
numpy>=1.24.0
psutil>=5.9.0
watchdog>=3.0.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
